strnatcmp.c: remove assert, remove unneeded headers
Remove the assert that checks if both input strings are non-NULL,
because the regular strcmp() does not check that either. Failing with a
segmentation fault is actually desirable here. A routine like this
should have no side effects, it should be as 'pure' as possible.

With the assert gone, we can remove the other header file inclusions, as
they no longer serve a purpose.